WIP Pulling in existing calibration functionality:  - Passing pose attributes to shader  - Created calibration panels and passing information to shader  - Started work on calibration loader class

+++ src/loader/BinaryLoader.js
@@ -96,9 +96,9 @@ export class BinaryLoader{
 						range: buffers[property].range,
 					};
 				} else if (parseInt(property) === PointAttributeNames.RTK_POSE) {
-					geometry.addAttribute('originalRtkPosition', new THREE.BufferAttribute(new Float64Array(buffer), 3, false));
+					geometry.addAttribute('originalRtkPosition', new THREE.BufferAttribute(new Float32Array(buffer), 3, false));
 				} else if (parseInt(property) === PointAttributeNames.RTK_ORIENT) {
-					geometry.addAttribute('originalRtkOrientation', new THREE.BufferAttribute(new Float64Array(buffer), 3, false));
+					geometry.addAttribute('originalRtkOrientation', new THREE.BufferAttribute(new Float32Array(buffer), 3, false));
 				}
 			}
